old/requirements.txt
pandas>=2.0
biopython>=1.81
scikit-learn>=1.2.2
numpy>=1.23.4
networkx>=3.1
scipy>=1.10.1
